DRF_Course/requirements.txt
active==0.0.5a0
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
flash-paymob==0.0.1
pytz==2022.1
sqlparse==0.4.2
